+++ src/main/java/Entities/PowerUp/IncreaseBallSpeed.java
@@ -6,7 +6,7 @@ public class IncreaseBallSpeed extends PowerUp  {
     private final Ball ball;
     private final double speedMultiplier = 1.5;
 
-    public IncreaseBallSpeed(float x, float y, Ball ball) {
+    public IncreaseBallSpeed(double x, double y, Ball ball) {
         super(x, y, 3, "/ibsPW.png");
         this.ball = ball;
         this.duration = 180; // e.g. 3 seconds at 60fps
